requirements.txt
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
moviepy==1.0.3
numpy==1.20.3
Pillow==8.2.0
proglog==0.1.9
requests==2.25.1
tqdm==4.60.0
urllib3==1.26.4
